backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
google-genai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
